Removed wrapper for dax dialog
